Migrating MoviesDetails to fragment
