+++ app/src/main/java/com/aacademy/homework/FragmentMoviesList.kt
@@ -1,0 +1,14 @@
+package com.aacademy.homework
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+class FragmentMoviesList : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_movies_list, container, false)
+    }
+}
