+++ app/src/main/res/layout/fragment_movies_details.xml
@@ -5,8 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/darkBlue"
-    android:fitsSystemWindows="false"
-    tools:context=".MovieDetailsActivity">
+    android:fitsSystemWindows="false">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
